src/components/form-control/json-preview-dialog/json-preview-dialog.ts
/**
 * @license
 * Copyright ASW (A Software World) All Rights Reserved.
 *
 * Use of this source code is governed by an MIT-style license that can be
 * found in the LICENSE file
 */

import { Component, Inject } from '@angular/core';
import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';

@Component({
    selector: 'asw-json-preview-dialog',
    templateUrl: './json-preview-dialog.html',
})
export class AswJsonPreviewDialog {
    constructor(
        public dialogRef: MatDialogRef<AswJsonPreviewDialog>,
        @Inject(MAT_DIALOG_DATA) public data: any) { }

    onNoClick(): void {
        this.dialogRef.close();
    }
}
